requirements.txt
python-json-logger==2.0.2
psutil==5.8.0
Pillow==7.2.0
opencv-contrib-python==4.5.4.60
pytesseract==0.3.4
scikit-image==0.17.2
numpy==1.19.0
deskew==0.10.39
boto3==1.17.57
PyYAML==6.0
opencv-python==4.3.0.36
pdfplumber==0.5.28
pdf2image==1.16.0
pdfreader==0.1.10
botocore==1.20.57
scipy==1.5.1
ray==1.9.0
arrow==1.2.2